black-hat-python/requirements.txt
PyGithub
pywin32
pycryptodome
